pyproject.toml
[tool.poetry]
name = "ass4"
version = "0.1.0"
description = ""
authors = ["suryaMadhav16 <rebbapragada.s@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
datasets = "^3.1.0"
langchain-pinecone = "^0.2.0"
langchain-openai = "^0.2.8"
langchain = "^0.3.7"
langchain-core = "^0.3.18"
langgraph = "^0.2.48"
semantic-router = "^0.0.72"
google-search-results = "^2.4.2"
jupyter = "^1.1.1"
cohere = "^5.11.3"
arxiv = "^2.1.3"
langchain-community = "^0.3.7"
openai = "^1.54.4"
langchain-anthropic = "^0.3.0"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
html2text = "^2024.2.26"
fastapi = "^0.115.5"
httpx = "^0.27.2"
langchain-groq = "^0.2.1"
langchain-google-genai = "^2.0.4"
langgraph-checkpoint = "^2.0.4"
langgraph-checkpoint-sqlite = "^2.0.1"
langsmith = "^0.1.143"
numexpr = "^2.10.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
pyowm = "^3.3.0"
setuptools = "^75.5.0"
streamlit = "^1.40.1"
duckduckgo-search = "^6.3.4"
langchain-aws = "^0.2.7"
serpapi = "^0.1.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.10"
markdown2 = "^2.5.1"
fpdf = "^1.7.2"
weasyprint = "^63.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
